Using React Router now able to switch between lists on Airtable. able to fetch and write data to Airtable.
